requirements.txt
Pillow>=10.0.0
cryptography>=41.0.0
pycolor>=1.0.0
termcolor>=2.3.0